Changed: updated dependencies and old package files moved to temporal 'old' folders
