+++ jest.config.js
@@ -1,10 +1,11 @@
 module.exports = {
   roots: ['<rootDir>/test'],
   transform: {
-    '^.+\\.ts$': 'ts-jest'
+    '^.+\\.ts$': 'ts-jest',
   },
   testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.ts$',
   moduleFileExtensions: ['ts', 'js', 'json', 'node'],
-  coveragePathIgnorePatterns: ['testHelpers']
-}
-  
+  coveragePathIgnorePatterns: ['testHelpers'],
+  // Temporarily ignore old tests
+  modulePathIgnorePatterns: ['<rootDir>/test/old'],
+};
