+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "js-dataverse",
-  "version": "1.0.48",
+  "version": "2.0.0",
   "description": "A Dataverse module for JavaScript/TypeScript",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
@@ -12,11 +12,9 @@
     "publish": "tsc && npm publish",
     "test": "jest --watchAll",
     "test:coverage": "jest --coverage --watchAll",
-    "test:ci": "jest --ci --maxWorkers=2 --detectOpenHandles --forceExit",
-    "eslint:check": "eslint --ext '.js,.jsx,.ts,.tsx' .",
-    "eslint:fix": "eslint --ext '.js,.jsx,.ts,.tsx' . --fix",
-    "format:check": "tsfmt --verify",
-    "format:fix": "tsfmt --replace"
+    "format": "prettier --write \"src/**/*.(js|ts)\"",
+    "lint": "eslint src --ext .js,.ts",
+    "lint:fix": "eslint src --fix --ext .js,.ts"
   },
   "repository": {
     "type": "git",
@@ -32,25 +30,14 @@
   },
   "homepage": "https://github.com/IQSS/dataverse-client-javascript#readme",
   "devDependencies": {
-    "@types/chai": "^4.2.1",
-    "@types/faker": "^4.1.5",
-    "@types/jest": "^24.0.18",
-    "@types/request-promise": "^4.1.44",
-    "@types/sinon": "^7.0.13",
-    "@typescript-eslint/eslint-plugin": "^2.1.0",
-    "@typescript-eslint/parser": "^2.1.0",
-    "chai": "^4.2.0",
-    "eslint": "^6.3.0",
-    "eslint-config-prettier": "^6.2.0",
-    "faker": "^4.1.0",
-    "jest": "^24.9.0",
-    "sinon": "^7.4.2",
-    "ts-jest": "^24.0.2",
-    "typescript": "^3.6.2",
-    "typescript-formatter": "^7.2.2"
-  },
-  "dependencies": {
-    "axios": "^0.21.1",
-    "request-promise": "^4.2.4"
+    "@types/jest": "29.4.0",
+    "@typescript-eslint/eslint-plugin": "5.54.0",
+    "@typescript-eslint/parser": "5.52.0",
+    "eslint": "8.35.0",
+    "eslint-plugin-jest": "27.2.1",
+    "jest": "29.4.3",
+    "prettier": "2.8.4",
+    "ts-jest": "29.0.5",
+    "typescript": "4.9.5"
   }
 }
