+++ src/old/@types/dataverseHeaders.ts
