+++ src/old/utils/metadataBlocksUtil.ts
